Fix type cast and update test command for legacy exceptions
Cast source offset to UInt32 in verbose exceptions mode in NuXJS.cpp to ensure type safety. Update buildAndTest.cmd to run tests with the '-s --legacy-exceptions' flags for NuXJS, improving test coverage for legacy exception handling.

+++ src/NuXJS.cpp
@@ -3212,7 +3212,7 @@ void Compiler::emit(Processor::Opcode opcode, Int32 operand) {
 		error(RANGE_ERROR, "Internal compiler limitations reached. Reduce code complexity.");
 	}
 #if (NUXJS_VERBOSE_EXCEPTIONS)
-	currentSection->emit(opcode, operand, p - sourceUnitBase);
+	currentSection->emit(opcode, operand, static_cast<UInt32>(p - sourceUnitBase));
 #else
 	currentSection->emit(opcode, operand);
 #endif
